管理者権限

+++ resources/views/join.blade.php
@@ -20,23 +20,29 @@
                 <p>{{$events->message}}</p>
                 <p>{{$events->others}}</p>
         </div>
-        
-            
-          
            @foreach($event_user2 as $event_user)
-           <div>{{$user2[$event_user->user_id-1]->name}}</div>
-           <div>{{$event_user->comment}}</div>
-           <a href='/{{$event_user->id}}/joinDelete'>削除する</a>
+               <div>{{Auth::user()::find($event_user->user_id)->name}}</div>
+               <div>{{$event_user->comment}}</div>
+               @if(Auth::user()->id == $events->user_id)
+                    <a href='/{{$event_user->id}}/joinDelete'>管理者削除する</a>
+               @endif
+               
+               @if(!(Auth::user()->id == $events->user_id) && Auth::user()->id == $event_user->user_id)
+                    <a href='/{{$event_user->id}}/joinDelete'>削除する</a>
+               @endif
+               
+               
            @endforeach
            
-            <form action="/candidate" method="POST">
-                @csrf
-                    <div><input type="hidden" name="event_users[user_id]" value="{{ Auth::user()->id }}"/></div>
-                    <div><input type="hidden" name="event_users[event_id]" value="{{$events->id}}"/></div>
-                    <div><textarea name="event_users[comment]" placeholder="参加にあたってのコメント"></textarea></div>
-                <input type="submit" value="参加する"/>
-            </form> 
-            
+            @if(!Auth::user()->joinedEvent($events->id))
+                <form action="/candidate" method="POST">
+                    @csrf
+                        <div><input type="hidden" name="event_users[user_id]" value="{{ Auth::user()->id }}"/></div>
+                        <div><input type="hidden" name="event_users[event_id]" value="{{$events->id}}"/></div>
+                        <div><textarea name="event_users[comment]" placeholder="参加にあたってのコメント"></textarea></div>
+                        <input type="submit" value="参加する"/>
+                </form> 
+            @endif
